Improve syntax highlighting for comments and diffs
